toolkit/content/tests/chrome/test_labelcontrol.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<!--
  XUL Widget Test for label control="value"
  -->
<window title="tabindex" width="500" height="600"
        onload="runTests()"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>  
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>  

<label id="textbox-label" control="textbox" />
<textbox id="textbox" value="Test"/>
<label id="checkbox-label" control="checkbox" />
<checkbox id="checkbox" value="Checkbox"/>

<body xmlns="http://www.w3.org/1999/xhtml">
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
</body>

<script>
<![CDATA[

SimpleTest.waitForExplicitFinish();
 
function runTests()
{
  let textbox = $("textbox");
  let textboxLabel = $("textbox-label");
  is(textboxLabel.control, "textbox", "textbox control");
  is(textboxLabel.labeledControlElement, textbox, "textbox labeledControlElement");

  let checkbox = $("checkbox");
  let checkboxLabel = $("checkbox-label");
  is(checkboxLabel.control, "checkbox", "checkbox control");
  is(checkboxLabel.labeledControlElement, checkbox, "checkbox labeledControlElement");
  is(checkbox.accessKey, "", "checkbox accessKey not set");
  checkboxLabel.accessKey = "C";
  is(checkbox.accessKey, "C", "checkbox accessKey set");

  SimpleTest.finish();
}

]]>

</script>

</window>
